Refactor schema for common structures
