+++ pkg/server/api/server/handler_test.go
@@ -1,0 +1,1 @@
+package server
