--- pkg/server/api/v1/handler_test.go
@@ -1,1 +1,0 @@
-package api
